+++ packages/conform-zod/v3/future.ts
@@ -1,4 +1,3 @@
 export { getZodConstraint } from './constraint';
 export { coerceFormValue } from './coercion';
 export { formatResult } from './format';
-export { memoize } from '../memoize';
